refactor: input, output view의 static 삭제
