+++ src/main/java/ui/OutputView.java
@@ -1,4 +1,4 @@
-package view;
+package ui;
 
 import java.util.List;
 import model.Cars;
@@ -6,15 +6,12 @@ import model.Cars;
 public class OutputView {
     private static final String WINNER_NAME_DELIMITER = ", ";
 
-    private OutputView() {
-    }
-
-    public static void printResultHeader() {
+    public void printResultHeader() {
         System.out.println();
         System.out.println("실행 결과");
     }
 
-    public static void printCarNameAndPosition(Cars cars) {
+    public void printCarNameAndPosition(Cars cars) {
         cars.getCars()
                 .forEach(car ->
                         System.out.printf("%s : %s%n", car.getName(), "-".repeat(car.getPosition()))
@@ -22,7 +19,7 @@ public class OutputView {
         System.out.println();
     }
 
-    public static void printWinners(List<String> winners) {
+    public void printWinners(List<String> winners) {
         System.out.println(String.join(WINNER_NAME_DELIMITER, winners) + "가 최종 우승했습니다.");
     }
 }
